refactor(client): Split PixiHexGrid.tsx into focused modules for LLM optimization
Extract rendering, camera, and utility code from 1697-line monolith into
focused modules optimized for LLM agents:

- pixi/pathfinding.ts: A* pathfinding for movement preview (108 lines)
- pixi/camera.ts: Pan/zoom controls with smooth interpolation (210 lines)
- pixi/rendering/tiles.ts: Tile rendering with intro animations (340 lines)
- pixi/rendering/enemies.ts: Enemy token rendering with drop effects (226 lines)
- pixi/rendering/hero.ts: Hero token with portal emergence (92 lines)
- pixi/rendering/overlays.ts: Movement highlights and cost badges (146 lines)
- pixi/rendering/pathPreview.ts: Path line with arrow visualization (86 lines)
- pixi/rendering/ghostHexes.ts: Exploration target indicators (87 lines)
- pixi/rendering/index.ts: Re-exports for convenient imports (12 lines)

Main component (PixiHexGrid.tsx) now focused on orchestration only (~560 lines).

Benefits for LLM agents:
- Focused files are easier to understand and modify
- Changes isolated to relevant modules reduce context needed
- Clear boundaries enable parallel work on different features
- Smaller diffs when making targeted changes
