+++ packages/client/src/components/GameBoard/pixi/rendering/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Rendering modules for PixiJS hex grid
+ *
+ * Re-exports all rendering functions for convenient imports.
+ */
+
+export { renderTiles, applyScreenShake } from "./tiles";
+export { renderEnemies } from "./enemies";
+export { renderHeroIntoContainer, getOrCreateHeroContainer } from "./hero";
+export { renderHexOverlays, type MoveHighlight, type MoveHighlightType } from "./overlays";
+export { renderPathPreview } from "./pathPreview";
+export { renderGhostHexes, type ExploreTarget } from "./ghostHexes";
